+++ src/lib/core/service.ts
@@ -4,7 +4,7 @@ import { Injectable } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { DropdownItem } from './../dropdown/dropdown-item.inteface';
 import { NotificationsService } from './../toast/simple-notifications/services/notifications.service';
-import { default as swal } from 'sweetalert2';
+import { default as swal, SweetAlertType } from 'sweetalert2';
 import { WizardConfig } from './wizard-config.interface';
 
 @Injectable()
@@ -15,15 +15,6 @@ export class Plex {
     public userInfo: any;
 
     constructor(private titleService: Title, private noficationService: NotificationsService) {
-    }
-
-    /**
-     * Inicializa la vista de la aplicación
-     * @deprecated Utilizar los métodos updateTitle() y updateMenu()
-     */
-    initView(title: string, menu: DropdownItem[] = null) {
-        this.titleService.setTitle(title);
-        this.menu = menu;
     }
 
     /**
@@ -80,16 +71,19 @@ export class Plex {
      * @memberof Plex
      * @deprecated Utilizar el método info()
      */
-    alert(content: string, title = 'Información'): Promise<any> {
+    alert(content: string, title = 'Información', confirmButtonText = 'ACEPTAR'): Promise<any> {
         return swal({
             title: title,
             html: content,
             type: 'warning',
-            confirmButtonText: 'Ok'
+            confirmButtonText: confirmButtonText.toLocaleUpperCase(),
+            buttonsStyling: false,
+            confirmButtonClass: 'btn btn-warning',
         });
     }
 
     /**
+     * TODO: Migrar para usar 1 sólo objeto con su type como param
      * Muestra un diálogo de confirmación
      *
      * @param {string} content Texto
@@ -98,44 +92,79 @@ export class Plex {
      *
      * @memberof Plex
      */
-    confirm(content: string, title = 'Confirmación'): Promise<any> {
-        let resolve: any;
-        let promise = new Promise((res, rej) => {
-            resolve = res;
-        });
+    confirm(params: { content: string, title: string, confirmButtonText: string, cancelButtonText: string });
 
-        swal({
-            title: title,
-            html: content,
-            type: 'question',
-            showCancelButton: true,
-            confirmButtonText: 'Confirmar',
-            cancelButtonText: 'Cancelar'
-        })
-            .then(() => resolve(true))
-            .catch(() => resolve(false));
-        return promise;
+    confirm(content: string, title?: string, confirmButtonText?: string, cancelButtonText?: string);
+
+    confirm(content, title = 'Confirmación', confirmButtonText = 'Confirmar', cancelButtonText = 'Cancelar'): Promise<any> {
+
+        let htmlContent;
+
+        // Para compatibilidad
+        if (typeof content === 'object') {
+            title = content.title || 'Confirmación';
+            htmlContent = content.content;
+            confirmButtonText = content.confirmButtonText || 'Confirmar';
+            cancelButtonText = content.cancelButtonText || 'Cancelar';
+        } else {
+            htmlContent = content;
+        }
+
+        return new Promise((resolve, reject) => {
+            swal({
+                title: title,
+                html: htmlContent,
+                type: 'question',
+                showCancelButton: true,
+                confirmButtonText: confirmButtonText.toLocaleUpperCase(),
+                cancelButtonText: cancelButtonText.toLocaleUpperCase(),
+                buttonsStyling: false,
+                confirmButtonClass: 'btn btn-success',
+                cancelButtonClass: 'btn btn-danger',
+            }).then(() => resolve(true))
+                .catch(() => resolve(false));
+        });
     }
 
     /**
+     * TODO: Migrar para usar 1 sólo objeto con su type como param
      * Muestra un mensaje invasivo al usuario
      *
-     * @param {string} type success, danger, warning, info
+     * @param {string} type success, danger (error), warning, info
      * @param {string} content Texto del mensaje
      * @param {string} [title='Información'] Título
      * @param {number} [timeOut=0] Tiempo en ms cuando se oculta el mensaje. Por default no se oculta.
      *
      * @memberof Plex
      */
-    info(type: string, content: string, title = 'Información', timeOut = 0) {
-        if (type === 'danger') {
-            type = 'error';
+    info(type: String, content: String, title?: String, timeOut?: Number, confirmButtonText?: String);
+    info(params: { type: String, content: String, title: String, confirmButtonText: String, timeOut?: Number });
+    info(type, content = '', title = 'Información', timeOut = 0, confirmButtonText = 'Aceptar') {
+        let modalType;
+
+        // Para compatibilidad
+        if (typeof type === 'object') {
+            // TODO: Usar el tipo SweetAlertType?
+            modalType = type.type === 'danger' ? 'error' : type.type;
+            content = type.content || '';
+            title = type.title || 'Información';
+            confirmButtonText = type.confirmButtonText ? type.confirmButtonText.toLocaleUpperCase() : 'Aceptar';
+            timeOut = type.timeOut || 0;
+        } else {
+            // TODO: Usar el tipo SweetAlertType?
+            if (type === 'danger') {
+                type = modalType = 'error';
+            }
+            modalType = type;
         }
+
         return swal({
             title: title,
             html: content,
-            type: type as any,
-            confirmButtonText: 'Ok',
+            type: modalType,
+            confirmButtonText: confirmButtonText,
+            buttonsStyling: false,
+            confirmButtonClass: `btn btn-${modalType === 'error' ? 'danger' : modalType}`,
             timer: timeOut || null,
         }).catch(swal.noop);
     }
@@ -150,7 +179,7 @@ export class Plex {
      *
      * @memberof Plex
      */
-    toast(type: string, content: string, title = 'Información', timeOut = 2500) {
+    toast(type: string, content: string, title: string = 'Información', timeOut: number = 2500) {
         let options = {
             theClass: 'toast',
             timeOut: timeOut
@@ -237,8 +266,15 @@ export class Plex {
                 });
             }
 
-            // Corrije los textos
+            // En el primer paso el botón principal dice "Comenzar"
             steps[0].confirmButtonText = 'Comenzar';
+
+            // En los pasos intermedios los botones dicen "Siguiente" y "Cancelar"
+            steps = steps.map(s => {
+                return { ...s, buttonsStyling: false, confirmButtonClass: 'btn btn-info', cancelButtonClass: 'btn btn-danger' };
+            });
+
+            // En el último paso el botón principal dice "Finalizar" y el botón "Cancelar" se oculta
             let last = steps[steps.length - 1];
             last.confirmButtonText = 'Finalizar';
             last.showCancelButton = false;
@@ -294,9 +330,11 @@ export class Plex {
                 .oncomplete(() => {
                     // No volver a mostrar
                     localStorage[`wizard-${config.id}-${config.updatedOn.toISOString()}-hide`] = true;
-                    resolve(true)
+                    resolve(true);
                 })
-                .onexit(() => resolve(false))
+                .onexit(
+                    () => resolve(false)
+                );
         }
 
         return promise;