+++ src/lib/datetime/datetime.component.ts
@@ -4,7 +4,9 @@ import * as moment from 'moment';
 import { dateValidator, hasRequiredValidator } from '../core/validator.functions';
 
 // Importo las librerías de jQuery
-let jQuery = window['jQuery'] = require('jquery/dist/jquery'); // @jgabriel: No encontré una forma más elegante de incluir jQuery
+// @jgabriel: No encontré una forma más elegante de incluir jQuery
+// @andrrr: qué mal
+let jQuery = window['jQuery'] = require('jquery/dist/jquery');
 require('./bootstrap-material-datetimepicker/bootstrap-material-datetimepicker');
 
 @Component({
@@ -23,14 +25,16 @@ require('./bootstrap-material-datetimepicker/bootstrap-material-datetimepicker')
         },
     ],
     template: ` <div class="form-group" [ngClass]="{'has-danger': (control.dirty || control.touched) && !control.valid }">
-                    <label *ngIf="label" class="form-control-label">{{label}}<span *ngIf="esOpcional" class="opcional"></span></label>
-                    <div class="input-group">
+                    <label *ngIf="label" class="form-control-label">{{ label }}</label>
+                    <div class="input-group d-flex align-items-center">
+                        <a *ngIf="showArraws" (click)="prev()" class="btn btn-info text-white pl-1 pr-1 hover" [title]="makeTooltip('anterior')"><i class="mdi mdi-menu-left"></i></a>
                         <input type="text" class="form-control" [placeholder]="placeholder" [disabled]="disabled" [readonly]="readonly" (input)="onChange($event.target.value)" (blur)="onBlur()"/>
                         <span class="input-group-btn">
-                            <button class="btn btn-primary" tabindex="-1" [disabled]="disabled || readonly"><i class="mdi" [ngClass]="{'mdi-calendar': type == 'date','mdi-clock': type == 'time', 'mdi-calendar-clock': type == 'datetime'}"></i></button>
+                        <button class="btn btn-primary" tabindex="-1" [disabled]="disabled || readonly"><i class="mdi" [ngClass]="{'mdi-calendar': type == 'date','mdi-clock': type == 'time', 'mdi-calendar-clock': type == 'datetime'}"></i></button>
                         </span>
+                        <a *ngIf="showArraws" (click)="next()" class="btn btn-info text-white pl-1 pr-1 hover" [title]="makeTooltip('siguiente')"><i class="mdi mdi-menu-right"></i></a>
                     </div>
-                    <plex-validation-messages *ngIf="(control.dirty || control.touched) && !control.valid" [control]="control"></plex-validation-messages>
+                    <plex-validation-messages *ngIf="hasDanger()" [control]="control"></plex-validation-messages>
                 </div>`,
 })
 export class PlexDateTimeComponent implements OnInit, AfterViewInit, OnChanges {
@@ -40,6 +44,7 @@ export class PlexDateTimeComponent implements OnInit, AfterViewInit, OnChanges {
     private value: any;
     private $button: any;
     private $input: any;
+
     @ContentChild(NgControl) control: AbstractControl;
     public get esOpcional(): boolean {
         return hasRequiredValidator(this.control);
@@ -52,6 +57,8 @@ export class PlexDateTimeComponent implements OnInit, AfterViewInit, OnChanges {
     @Input() placeholder: string;
     @Input() disabled = false;
     @Input() readonly = false;
+    @Input() skipBy: 'hour' | 'day' | 'month' | 'year' = null;
+    @Input() title: string;
     @Input()
     get min(): Date | moment.Moment {
         return this._min;
@@ -77,6 +84,10 @@ export class PlexDateTimeComponent implements OnInit, AfterViewInit, OnChanges {
                 this.$button.bootstrapMaterialDatePicker('setMaxDate', this._max);
             }
         }
+    }
+
+    get showArraws(): Boolean {
+        return this.skipBy && this.value;
     }
 
     // Eventos
@@ -105,15 +116,15 @@ export class PlexDateTimeComponent implements OnInit, AfterViewInit, OnChanges {
     // Inicialización
     ngOnInit() { }
     ngAfterViewInit() {
-        this.format = this.type === 'date' ? 'DD/MM/YYYY' : (this.type === 'datetime' ? 'DD/MM/YYYY HH:mm' : 'HH:mm');
+        this.format = this.dateOrTime();
         this.$input = jQuery('INPUT', this.element.nativeElement.children[0]);
         this.$button = jQuery('BUTTON', this.element.nativeElement.children[0]);
         this.$button.bootstrapMaterialDatePicker({
             lang: 'es',
             format: this.format,
             currentDate: this.value,
-            okText: 'Ok',
-            cancelText: 'Cancelar',
+            okText: 'ACEPTAR',
+            cancelText: 'CANCELAR',
             clearButton: false,
             nowButton: false,
             switchOnClick: true,
@@ -128,22 +139,25 @@ export class PlexDateTimeComponent implements OnInit, AfterViewInit, OnChanges {
         });
     }
 
+    private dateOrTime(): string {
+        return this.type === 'date' ? 'DD/MM/YYYY' : (this.type === 'datetime' ? 'DD/MM/YYYY HH:mm' : 'HH:mm');
+    }
+
     // Actualización Modelo -> Vista
     writeValue(value: any) {
         this.value = value;
-
         let temp = this.value ? moment(this.value).format(this.format) : null;
-        if (this.$button) {
-            this.$button.val(temp);
-        }
-        if (this.$input) {
-            this.$input.val(temp);
-        }
+        this.setElements(temp);
+    }
+
+    hasDanger() {
+        return (this.control as any).name && (this.control.dirty || this.control.touched) && !this.control.valid;
     }
 
     // Actualización Vista -> Modelo
     registerOnTouched() {
     }
+
     registerOnChange(fn: any) {
         this.onChange = (value) => {
             if (typeof value === 'string') {
@@ -177,5 +191,31 @@ export class PlexDateTimeComponent implements OnInit, AfterViewInit, OnChanges {
                 return (fecha1 && fecha2 && fecha1.getTime() !== fecha2.getTime());
             }
         }
+    }
+
+    prev() {
+        let temp = this.value ? moment(this.value, this.dateOrTime()).subtract(1, this.skipBy).format(this.format) : null;
+        this.setElements(temp);
+        this.onChange(this.value);
+    }
+
+    next() {
+        let temp = this.value ? moment(this.value, this.dateOrTime()).add(1, this.skipBy).format(this.format) : null;
+        this.setElements(temp);
+        this.onChange(this.value);
+    }
+
+    private setElements(temp: string) {
+        if (this.$button) {
+            this.$button.val(temp);
+        }
+        if (this.$input) {
+            this.$input.val(temp);
+        }
+        this.value = temp;
+    }
+
+    makeTooltip(dir) {
+        return this.skipBy === 'hour' ? `hora ${dir}` : this.skipBy === 'day' ? `día ${dir}` : this.skipBy === 'month' ? `mes ${dir}` : `año ${dir}`;
     }
 }