+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [3.6.1] - 2018-09-19
+
+#### Added
+
+* plex.alert again.
+
+## [3.6.0] - 2018-09-19
+
+#### Added
+
+* Propertie skipBy on plex-datetime.
+
 ## [3.2.0] - 2018-06-11
 
 #### Added